add db, comments

+++ app/Http/Controllers/JobController.php
@@ -10,20 +10,29 @@ use Illuminate\Support\Facades\Auth;
 class JobController extends Controller
 {
     //
+    // function return homepage with all job posts from database
     public function homepage()
     {
         $jobs = Job::all();
         return view('homepage', ['jobs' => $jobs]);
     }
 
+    // Function returns page with form for creating Job post
     public function createJobForm()
     {
         return view('createJobForm');
     }
 
+
+    // Function creates new Job post and insert it in DB
     public function createJobAd(Request $request)
     {
 
+        if (!Auth::check()) {
+            return redirect()->route('homepage');
+        }
+
+        // validating data from request
         $request->validate([
             'job_position' => "required",
             'company' => "required",
@@ -32,13 +41,16 @@ class JobController extends Controller
             'salary' => "required",
         ]);
 
+        // After validation creating new job post
         $jobAd = new Job();
         $jobAd->job_position = $request->job_position;
         $jobAd->company = $request->company;
         $jobAd->location = $request->location;
         $jobAd->description = $request->description;
         $jobAd->salary = $request->salary;
+        // Logged in user id
         $jobAd->user_id = Auth::user()->id;
+        // Inserting in DB
         $jobAd->save();
         return redirect(route('homepage'));
     }
@@ -53,8 +65,11 @@ class JobController extends Controller
     public function deleteJob($id)
     {
         $job = Job::findOrFail($id);
-        $job->delete();
+        if (Auth::user()->id === $job->user_id) {
+            $job->delete();
+            return redirect()->route('homepage')->with('success', 'Job post deleted successfully');
+        }
 
-        return redirect()->route('homepage')->with('success', 'Job post deleted successfully');
+        return redirect()->route('homepage');
     }
 }
